requirements.txt
Flask==2.3.2
requests==2.31.0
cachetools==5.3.1
python-dotenv==1.0.0
Flask-CORS==3.0.10